move Vector.toJson/fromJson to ml.linalg.VectorEncoder under mllib
